test: ラインチャート用エンドポイントとその作成
